Remove Spellcaster component and inline behavior into AI class
