+++ pyro/ai/aggressive_spellcaster.py
@@ -1,6 +1,6 @@
 import tcod as libtcod
 import pyro.astar
-from pyro.components import AI, Spellcaster
+from pyro.components import AI
 from pyro.spell import SpellType
 from pyro.engine.actions import WalkAction
 
@@ -12,19 +12,19 @@ class AggressiveSpellcaster(AI):
         if monster.game.map.is_in_fov(monster.pos.x, monster.pos.y):
             # Heal yourself if damaged
             if monster.actor.hp < monster.actor.max_hp:
-                heals = self.owner.component(Spellcaster).get_spells(SpellType.HEAL)
+                heals = self.get_spells(SpellType.HEAL)
                 if len(heals) > 0:
-                    self.owner.component(Spellcaster).cast_spell(action, heals[0], self.owner)
+                    self.cast_spell(action, heals[0], self.owner)
                     return
 
             # Move towards player if far away
-            if not monster.component(Spellcaster).in_range(player, SpellType.ATTACK):
+            if not self.in_range(player, SpellType.ATTACK):
                 direction = pyro.astar.astar(self.owner.game, monster.pos, player.pos)
                 return WalkAction(direction)
 
             # Close enough, attack! (If the player is still alive)
             elif player.actor.hp > 0:
-                attacks = self.owner.component(Spellcaster).get_spells(SpellType.ATTACK)
+                attacks = self.get_spells(SpellType.ATTACK)
                 if len(attacks) > 0:
                     random_attack = attacks[libtcod.random_get_int(0, 0, len(attacks)-1)]
-                    self.owner.component(Spellcaster).cast_spell(action, random_attack, player)
+                    self.cast_spell(action, random_attack, player)
