+++ pyro/components/__init__.py
@@ -3,4 +3,3 @@ from ai import AI
 from graphics import Graphics
 from experience import Experience
 from item import Item, Inventory, Equipment, ItemUse, SpellItemUse
-from spellcaster import Spellcaster
